Implement infraestructure setup and initial migrations

+++ OrderService/OrderService.Domain/Entities/Product.cs
@@ -5,5 +5,5 @@ public class Product
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public decimal UnitPrice { get; set; }
-    public decimal AvailableStock { get; set; }
+    public int AvailableStock { get; set; }
 }
